// src/app/page/master/kelompok-user-aplikasi/kelompok-user-aplikasi.interface.ts
import {EventEmitter} from '@angular/core';

export interface KelompokUserAplikasi {
	kdProfile?: string;
	kdAgama?: string;
	namaAgama?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
	
}